convert permissions to promises,

+++ api-v1/paths/administration/{user_id}/{permission_id}.js
@@ -14,7 +14,7 @@ module.exports = function(toolKit, userService, user2PermissionService) {
 				user2PermissionService.updateUserPermission(user_id, permission_id, details, () => {
 					res.status(200).json(toolKit.createSimpleResponse("success", "permission updated"))
 				});
-				
+
 			} else {
 				res.status(401).json(toolKit.createSimpleResponse("error", 'user does not have administrative privileges'));
 			}
@@ -65,12 +65,12 @@ module.exports = function(toolKit, userService, user2PermissionService) {
 			var user_id = req.params.user_id;
 			var permission_id = req.params.permission_id;
 			var details = req.query.details ? req.query.details : "";
-			
+
 			if (user2PermissionService.validatePermission('sys_administrator', req.session.permissions)) {
 				user2PermissionService.addUserPermission(user_id, permission_id, details, () => {
 					res.status(200).json(toolKit.createSimpleResponse("success", "permission deleted"))
 				});
-				
+
 			} else {
 				res.status(401).json(toolKit.createSimpleResponse("error", 'user does not have administrative privileges'));
 			}
@@ -118,20 +118,31 @@ module.exports = function(toolKit, userService, user2PermissionService) {
 			}
 		}
 	};
-	
+
 	function DELETE(req, res, next) {
 		try {
 			var user_id = req.params.user_id;
 			var permission_id = req.params.permission_id;
-			
+
+
 			if (user2PermissionService.validatePermission('sys_administrator', req.session.permissions)) {
 				if (user_id == 'admin') {
-					res.status(401).json(toolKit.createSimpleResponse("error", "administrator is blocked from deleting its own permissions"));
+					var permissionObj = req.session.permissions.find(o => o.permission_id === permission_id)
+					if (permissionObj.tags.indexOf("SYSTEM")>-1) {
+						res.status(401).json(toolKit.createSimpleResponse("error", "administrator cannot delete its SYSTEM privileges"));
+					} else {
+						user2PermissionService.deleteUserPermission(user_id, permission_id, () => {
+							res.status(200).json(toolKit.createSimpleResponse("success", "permission deleted"))
+						});
+					}
+
+				} else {
+					user2PermissionService.deleteUserPermission(user_id, permission_id, () => {
+						res.status(200).json(toolKit.createSimpleResponse("success", "permission deleted"))
+					});
 				}
-				user2PermissionService.deleteUserPermission(user_id, permission_id, () => {
-					res.status(200).json(toolKit.createSimpleResponse("success", "permission deleted"))
-				});
-				
+
+
 			} else {
 				res.status(401).json(toolKit.createSimpleResponse("error", 'user does not have administrative privileges'));
 			}
